Aleksei Krutskikh: 2026-04-25: Package docmcp source modules
