+++ requirements-dev.txt
@@ -1,3 +1,5 @@
 # Dependency auditing
 build>=1.2,<2
 pip-audit>=2.7,<3
+setuptools>=68
+wheel>=0.42
