+++ tech-debt/2026-04-25.md
@@ -1,0 +1,13 @@
+# Tech Debt
+
+Deferred implementation items captured during doc_mcp work.
+
+## TD-20260425-01
+- Title: Crawl and index PDF documents like HTML pages
+- Context: The current Playwright crawler indexes rendered HTML pages, but PDF links discovered during traversal are treated as static assets and skipped instead of being fetched, converted to text or Markdown, and stored in the page index.
+- Location: `src/docmcp/crawl_cli.py`, `src/docmcp/index_store.py`, `documentation/crawling.md`
+- Impact/Risk: Documentation sites that publish reference material, release notes, API docs, or policy documents as PDFs will have incomplete search and fetch coverage even when those PDFs are linked from crawled HTML pages.
+- Why Deferred Now: The current task is to record the crawler limitation as technical debt; implementing PDF support requires URL classification changes, PDF download/parsing dependencies, content-type handling, metadata decisions, and tests.
+- Proposed Fix: Stop classifying PDFs as excluded static assets, add PDF response detection and parsing, convert extracted PDF text into indexable Markdown, preserve the source URL/title metadata, and document the supported behavior and dependency requirements.
+- Acceptance Signal: A crawl of a site containing linked PDFs stores those PDF URLs in the SQLite index with searchable extracted content, while existing HTML crawl behavior and static asset filtering remain covered by tests.
+- Status: open
